- move multimodal into fmt_conversion
