+++ tools/fmt_conversion/multimodal/absolute_path_to_relative_path.py
