+++ tools/fmt_conversion/multimodal/source_format_to_data_juicer_format/video_chatgpt_to_dj.py
@@ -37,8 +37,8 @@ from loguru import logger
 from tqdm import tqdm
 
 from data_juicer.utils.mm_utils import SpecialTokens
-from tools.multimodal.utils import (check_args_load_to_dj_data,
-                                    convert_text_to_dj)
+from tools.fmt_conversion.multimodal.utils import (check_args_load_to_dj_data,
+                                                   convert_text_to_dj)
 
 
 @logger.catch(reraise=True)
